refactor(balances): extract functions that can be imported in browser context
